--- a/README.md
+++ b/README.md
@@ -7,17 +7,20 @@
 - If the program crashes, create an issue with the error.
 
 ## INSTALLATION:
-1. Install Python 3.10.6 from https://www.python.org/downloads/windows/
-   - Note that Python 3.10.6 cannot be used on Windows 7 or earlier.
+1. Install Python 3.11.4 from https://www.python.org/downloads/windows/
+   - Note that Python 3.11.4 cannot be used on Windows 7 or earlier.
 2. Clone the repository or download it from here https://github.com/jfd02/TFT-OCR-BOT/archive/refs/heads/main.zip
 3. Open Command Prompt and change the current directory to the folder where main.py is located 
-4. Run pip install -r requirements.txt in Command Prompt
-5. Install tesseract using the Windows installer available at: https://github.com/UB-Mannheim/tesseract/wiki
+4. Run 'pip install -r requirements.txt' in Command Prompt
+5. Install tesseract 5.3.1.20230401 using the Windows installer available at: https://github.com/UB-Mannheim/tesseract/wiki
    - Note the tesseract path from the installation.
-   - Set the tesseract path in the settings.py file (it may already be correct)
-6. Configure settings.py so the league client path is correct
-7. Disable all in-game overlays
-8. Run the main.py file
+   - Set Tesseract tessdata folder path in settings.py file (probably already correct)
+6. Download tesserocr v2.6.0 at: https://github.com/simonflueckiger/tesserocr-windows_build/releases
+7. Open Command Prompt and change the current directory to the folder where tesserocr-2.6.0-cp311-cp311-win_amd64.whl is located
+8. Run 'pip install tesserocr-2.6.0-cp311-cp311-win_amd64.whl' in Command Prompt
+7. Configure settings.py so the league client path is correct
+8. Disable all in-game overlays
+9. Run the main.py file
 
 ## FEATURES:
 ![main](https://i.imgur.com/1bXOmag.png)