--- a/requirements.txt
+++ b/requirements.txt
@@ -1,9 +1,9 @@
-numpy==1.23.0
-opencv_python==4.6.0.66
-Pillow==9.3.0
+numpy==1.25.2
+opencv_python==4.8.0.74
+Pillow==10.0.0
 PyDirectInput==1.0.4
-pytesseract==0.3.9
-requests==2.28.1
-screeninfo==0.8
-urllib3==1.26.9
+pytesseract==0.3.10
+requests==2.31.0
+screeninfo==0.8.1
+urllib3==2.0.4
 pypiwin32==223