--- a/requirements.txt
+++ b/requirements.txt
@@ -2,7 +2,6 @@ numpy==1.25.2
 opencv_python==4.8.0.74
 Pillow==10.0.0
 PyDirectInput==1.0.4
-pytesseract==0.3.10
 requests==2.31.0
 screeninfo==0.8.1
 urllib3==2.0.4