--- a/settings.py
+++ b/settings.py
@@ -2,5 +2,5 @@
 
 FORFEIT = False
 FORFEIT_TIME = 600  # Time in seconds
-LEAGUE_CLIENT_PATH = None # Replace None with your path. Ex: 'C:\\Riot Games\\League of Legends'
-TESSERACT_PATH = r'C:\\Program Files\\Tesseract-OCR\\tesseract.exe'
+LEAGUE_CLIENT_PATH = r'C:\\Riot Games\\League of Legends' # Replace None with your path. Ex: 'C:\\Riot Games\\League of Legends'
+TESSERACT_TESSDATA_PATH = r'C:\\Program Files\\Tesseract-OCR\\tessdata'