1. Change pytesseract to tesserocr
2. Update python to 3.11.4
3. Update requirements
4. Update readme

Switching from pytesseract to using tesserocr would result in a significant performance improvement. On my old laptop with an i5-7300HQ processor, the time required to detect a single item and all item has been reduced from 0.48 seconds and 6 seconds to 0.12 seconds and 2.2 seconds respectively. This means that the performance of running tesserocr on an i5-7300HQ is now equivalent to running pytesseract on an i7-12700k.

Furthermore, there is a high probability of encountering errors when using "pip install tesserocr" directly. Therefore, I choose to first download the .whl file and then install it.

i5-7300hq with tesserocr (0.48s->0.12s)
![image](https://github.com/jfd02/TFT-OCR-BOT/assets/36875936/0e576ba7-26eb-4acd-85e1-0527fe417439)

i7-12700k with tesserocr (0.11s->0.06s)
![image](https://github.com/jfd02/TFT-OCR-BOT/assets/36875936/aa1ced16-2af5-4fb8-8ada-a911ffff256d)

#218